separate tests from runtime impl
This will help for easier review once the refactoring begins.
